Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -16,7 +16,7 @@ int linear_search(int *array, size_t size, int value)
 	if (array == NULL)
 		return (-1);
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
 		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (value == array[i])
